simplified the project to contain two code files, two parameters, a single model
