+++ get-started-experiments/code-experiments/params.yaml
@@ -1,5 +1,5 @@
-epochs: 10
-conv_units: 16
-dense_units: 128
-dropout: 0.5
+train:
+  epochs: 10
+model: 
+  conv_units: 16
